Backend/requirements.txt
Django==5.2.6
djangorestframework==3.15.2
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
Pillow==10.4.0


